backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==5.2.10
django-celery-beat==2.8.1
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
django-timezone-field==7.2.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
google-generativeai==0.3.0
kombu==5.6.2
packaging==26.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.5.3
